requirements.txt
langchain 
wikipedia 
langchain-community 
python-dotenv
pydantic
duckduckgo-search
langchain-huggingface
huggingface_hub